+++ config/medium/Squares/code/CGCode.cpp
@@ -1,0 +1,30 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    int side;
+    int m;
+    int n;
+    cin >> side >> m >> n; cin.ignore();
+    for (int i = 0; i < m; i++) {
+        int x;
+        int y;
+        int s;
+        cin >> x >> y >> s; cin.ignore();
+    }
+    for (int i = 0; i < n; i++) {
+        int tx;
+        int ty;
+        cin >> tx >> ty; cin.ignore();
+    }
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "answer" << endl;
+}