+++ config/medium/Squares/code/CGCode.js
@@ -1,0 +1,20 @@
+var inputs = readline().split(' ');
+const side = parseInt(inputs[0]);
+const M = parseInt(inputs[1]);
+const N = parseInt(inputs[2]);
+for (let i = 0; i < M; i++) {
+    var inputs = readline().split(' ');
+    const x = parseInt(inputs[0]);
+    const y = parseInt(inputs[1]);
+    const s = parseInt(inputs[2]);
+}
+for (let i = 0; i < N; i++) {
+    var inputs = readline().split(' ');
+    const tx = parseInt(inputs[0]);
+    const ty = parseInt(inputs[1]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');