+++ config/medium/Squares/code/CGCode.php
@@ -1,0 +1,15 @@
+<?php
+fscanf(STDIN, "%d %d %d", $side, $M, $N);
+for ($i = 0; $i < $M; $i++)
+{
+    fscanf(STDIN, "%d %d %d", $x, $y, $s);
+}
+for ($i = 0; $i < $N; $i++)
+{
+    fscanf(STDIN, "%d %d", $tx, $ty);
+}
+
+// Write an answer using echo(). DON'T FORGET THE TRAILING \n
+// To debug: error_log(var_export($var, true)); (equivalent to var_dump)
+
+echo("answer\n");