+++ config/medium/Squares/code/CGCode.ts
@@ -1,0 +1,20 @@
+var inputs: string[] = readline().split(' ');
+const side: number = parseInt(inputs[0]);
+const M: number = parseInt(inputs[1]);
+const N: number = parseInt(inputs[2]);
+for (let i = 0; i < M; i++) {
+    var inputs: string[] = readline().split(' ');
+    const x: number = parseInt(inputs[0]);
+    const y: number = parseInt(inputs[1]);
+    const s: number = parseInt(inputs[2]);
+}
+for (let i = 0; i < N; i++) {
+    var inputs: string[] = readline().split(' ');
+    const tx: number = parseInt(inputs[0]);
+    const ty: number = parseInt(inputs[1]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');